Rename source package to edu.sjsu.stratagem
